RandomKrylov

Documentation for RandomKrylov.

RandomKrylov.RandomVectorSketch

RandomKrylov.RandomizedBlockKaczmarz

RandomKrylov.RandomizedCD

RandomKrylov.RandomizedCD2

RandomKrylov.RandomizedKaczmarz

RandomKrylov.RandomizedNewton

RandomKrylov.randomInv

RandomKrylov.random_matrix_1

RandomKrylov.random_matrix_2

RandomKrylov.random_matrix_3

RandomKrylov.random_matrix_4

RandomKrylov.random_projector_rank

RandomKrylov.random_projector_sparse

RandomKrylov.RandomVectorSketch — Method

Random vector sketch

Section 3.2 in Gower, R. M., & Richtárik, P. (2015). Randomized iterative methods for linear systems. SIAM Journal on Matrix Analysis and Applications, 36(4), 1660-1690.

RandomKrylov.RandomizedBlockKaczmarz — Method

Randomized block Kaczmarz

Section 3.5 in Gower, R. M., & Richtárik, P. (2015). Randomized iterative methods for linear systems. SIAM Journal on Matrix Analysis and Applications, 36(4), 1660-1690.

RandomKrylov.RandomizedCD — Method

Randomized coordinate descent

Section 3.7 in Gower, R. M., & Richtárik, P. (2015). Randomized iterative methods for linear systems. SIAM Journal on Matrix Analysis and Applications, 36(4), 1660-1690.

RandomKrylov.RandomizedCD2 — Method

Randomized coordinate descent for symmetric positive definite matrix

Section 3.4 in Gower, R. M., & Richtárik, P. (2015). Randomized iterative methods for linear systems. SIAM Journal on Matrix Analysis and Applications, 36(4), 1660-1690.

RandomKrylov.RandomizedKaczmarz — Method

Randomized Kaczmarz

Section 3.3 in Gower, R. M., & Richtárik, P. (2015). Randomized iterative methods for linear systems. SIAM Journal on Matrix Analysis and Applications, 36(4), 1660-1690.

RK takes a step in the direction of the negative stochastic gradient. This means that it is equivalent to the SGD method. However, the stepsize choice is very special: RK chooses the stepsize which leads to the point which is closest to x* in the Euclidean norm.

RandomKrylov.RandomizedNewton — Method

Randomized Newton -> for symmetric positive definite matrix

Section 3.6 in Gower, R. M., & Richtárik, P. (2015). Randomized iterative methods for linear systems. SIAM Journal on Matrix Analysis and Applications, 36(4), 1660-1690.

RandomKrylov.randomInv — Method

Solve TAx=Tb with T a random projector

RandomKrylov.random_matrix_1 — Method

Random projector of size k x m with a normal distribution

RandomKrylov.random_matrix_2 — Method

Random projector of size k x m with -1 or 1 both with probability 1/2

RandomKrylov.random_matrix_3 — Method

Random projector of size k x m with -1,0,1 respectively with probability 1/6,4/6,1/6

RandomKrylov.random_matrix_4 — Method

Random projector of size k x m with orthogonal projection on a random k-dimensional linear subspace of R^m

RandomKrylov.random_projector_rank — Method

Check the average rank of random projector over N random matrices of size k*m

RandomKrylov.random_projector_sparse — Method

Check the average sparsity of random projector over N random matrices of size k*m